refactor(index): extract proximadb-index-types foundation crate (decomp contracts step 1)
First implementation step of the storage<->index/compute decoupling contracts
(docs/12-design/STORAGE_INDEX_COMPUTE_DECOUPLING_CONTRACTS_2026_06_21.adoc): move the
pure AXIS config/algorithm types (IndexAlgorithm, AxisConfig, Data, IndexSpecification,
IndexSelectionStrategy, ClusterAssignment, migration/monitoring configs — src/index/axis/types.rs,
615 LOC, serde-only, zero code deps) into a clean foundation crate.

- New crate crates/foundation/proximadb-index-types (serde only).
- src/index/axis/types.rs becomes a re-export shim (`pub use proximadb_index_types::*`),
  so all 28 `crate::index::axis::types::*` dependents + types_tests are unchanged.
- Two rustdoc intra-doc links to a sibling module converted to plain text (would dangle
  in the new crate).

Pure mechanical move, cold subtree (src/index untouched by active sessions), no behavioural
change. Verified: new crate builds (serde only); full `cargo check -p proximadb --lib` green.
This is the low-risk prerequisite that unblocks decoupling storage from the index types layer.
