+++ crates/foundation/proximadb-index-types/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "proximadb-index-types"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "Pure AXIS index configuration/algorithm types (IndexAlgorithm, AxisConfig, IndexSpecification, ...) — foundation layer shared by storage, index, and query planning"
+
+[lib]
+name = "proximadb_index_types"
+path = "src/lib.rs"
+
+[dependencies]
+serde = { workspace = true, features = ["derive"] }
+
+[dev-dependencies]
+serde_json.workspace = true
